split parser modules a bit more
