+++ crates/hcl-edit/src/parser/expr.rs
@@ -1,10 +1,13 @@
+use super::ident::{ident, str_ident};
 use super::number::number;
+use super::repr::{decor, prefix_decor, spanned, suffix_decor};
+use super::string::{raw, string};
+use super::trivia::{line_comment, sp, ws};
 use super::{
-    cut_char, cut_ident, cut_tag, decor,
+    cut_char, cut_ident, cut_tag,
     error::{Context, Expected, InternalError},
-    ident, line_comment, prefix_decor, raw, sp, spanned, str_ident, string, suffix_decor,
     template::{heredoc_template, string_template},
-    ws, IResult, Input,
+    IResult, Input,
 };
 use crate::expr::{
     Array, BinaryOp, BinaryOperator, Conditional, Expression, ForCond, ForExpr, ForIntro, FuncCall,
